Move the test-only legacy column bridges into test support, guard the scroll retry, and refresh the store perf docs
